update  component
