+++ components/Modals/BuyNFT/index.tsx
@@ -449,8 +449,8 @@ function Component() {
                             />
                         </div>
                         {txType === 'AddFunds' &&
-                            (loginInfo.txStatusLoading === 'true' ||
-                                loginInfo.txStatusLoading === 'submitted') ? (
+                        (loginInfo.txStatusLoading === 'true' ||
+                            loginInfo.txStatusLoading === 'submitted') ? (
                             <div className={styles.wrapperLoading}>
                                 <div className={styles.loadingIco}>
                                     {spinner}
@@ -470,9 +470,9 @@ function Component() {
                                     <h2 className={styles.usernameInfoYellow}>
                                         {username?.length! > 20
                                             ? `${username?.slice(
-                                                0,
-                                                8
-                                            )}...${username?.slice(-8)}`
+                                                  0,
+                                                  8
+                                              )}...${username?.slice(-8)}`
                                             : username}
                                     </h2>
                                     <h2 className={styles.usernameInfo}>
@@ -641,9 +641,9 @@ function Component() {
                                             >
                                                 {buyInfo?.recipientOpt ===
                                                     'SSI' ||
-                                                    (buyInfo?.recipientOpt ===
-                                                        'ADDR' &&
-                                                        buyInfo?.anotherAddr !==
+                                                (buyInfo?.recipientOpt ===
+                                                    'ADDR' &&
+                                                    buyInfo?.anotherAddr !==
                                                         undefined) ? (
                                                     <>
                                                         <div
@@ -696,9 +696,9 @@ function Component() {
                                             >
                                                 {buyInfo?.recipientOpt ===
                                                     'SSI' ||
-                                                    (buyInfo?.recipientOpt ===
-                                                        'ADDR' &&
-                                                        buyInfo?.anotherAddr !==
+                                                (buyInfo?.recipientOpt ===
+                                                    'ADDR' &&
+                                                    buyInfo?.anotherAddr !==
                                                         undefined) ? (
                                                     <>
                                                         <div
@@ -726,7 +726,7 @@ function Component() {
                                         </div>
                                         {buyInfo?.recipientOpt == 'ADDR' ? (
                                             buyInfo?.anotherAddr !==
-                                                undefined ? (
+                                            undefined ? (
                                                 <p style={{ marginTop: '3%' }}>
                                                     {t('Recipient (address):')}{' '}
                                                     {zcrypto.toBech32Address(
@@ -767,7 +767,7 @@ function Component() {
                                                         <div
                                                             className={
                                                                 legend ===
-                                                                    'save'
+                                                                'save'
                                                                     ? 'continueBtn'
                                                                     : ''
                                                             }
@@ -776,7 +776,7 @@ function Component() {
                                                             }
                                                         >
                                                             {legend ===
-                                                                'save' ? (
+                                                            'save' ? (
                                                                 <Image
                                                                     src={
                                                                         ContinueArrow
@@ -812,46 +812,46 @@ function Component() {
                                             <>
                                                 {buyInfo?.currency !==
                                                     'FREE' && (
-                                                        <div
+                                                    <div
+                                                        className={
+                                                            styles.balanceInfoWrapepr
+                                                        }
+                                                    >
+                                                        <p
                                                             className={
-                                                                styles.balanceInfoWrapepr
+                                                                styles.balanceInfo
                                                             }
                                                         >
+                                                            {t(
+                                                                'CURRENT_BALANCE'
+                                                            )}
+                                                        </p>
+                                                        {loadingBalance ? (
+                                                            <div
+                                                                style={{
+                                                                    marginLeft:
+                                                                        '2%',
+                                                                }}
+                                                            >
+                                                                {spinner}
+                                                            </div>
+                                                        ) : (
                                                             <p
                                                                 className={
-                                                                    styles.balanceInfo
+                                                                    styles.balanceInfoYellow
                                                                 }
                                                             >
-                                                                {t(
-                                                                    'CURRENT_BALANCE'
-                                                                )}
+                                                                &nbsp;
+                                                                {
+                                                                    buyInfo?.currentBalance
+                                                                }{' '}
+                                                                {
+                                                                    buyInfo?.currency
+                                                                }
                                                             </p>
-                                                            {loadingBalance ? (
-                                                                <div
-                                                                    style={{
-                                                                        marginLeft:
-                                                                            '2%',
-                                                                    }}
-                                                                >
-                                                                    {spinner}
-                                                                </div>
-                                                            ) : (
-                                                                <p
-                                                                    className={
-                                                                        styles.balanceInfoYellow
-                                                                    }
-                                                                >
-                                                                    &nbsp;
-                                                                    {
-                                                                        buyInfo?.currentBalance
-                                                                    }{' '}
-                                                                    {
-                                                                        buyInfo?.currency
-                                                                    }
-                                                                </p>
-                                                            )}
-                                                        </div>
-                                                    )}
+                                                        )}
+                                                    </div>
+                                                )}
                                                 {buyInfo?.currency !==
                                                     undefined &&
                                                     !loadingBalance && (
@@ -859,7 +859,7 @@ function Component() {
                                                             {buyInfo?.isEnough ? (
                                                                 <>
                                                                     {donation ===
-                                                                        null ? (
+                                                                    null ? (
                                                                         <Donate />
                                                                     ) : (
                                                                         <>
@@ -885,8 +885,8 @@ function Component() {
                                                                                     {loading
                                                                                         ? spinner
                                                                                         : t(
-                                                                                            'BUY NFT USERNAME'
-                                                                                        )}
+                                                                                              'BUY NFT USERNAME'
+                                                                                          )}
                                                                                 </div>
                                                                             </div>
                                                                             <h5
